+++ lib/catechism.dart
@@ -5,3 +5,4 @@ export 'src/application.dart';
 export 'src/configuration.dart';
 export 'src/models/info_screen_data.dart';
 export 'src/models/question.dart';
+export 'src/models/tab.dart';