+++ lib/src/models/question.dart
@@ -7,11 +7,15 @@ class Question with ChangeNotifier {
   final String title;
   final String question;
   final String answer;
+  final List<String> references;
+  final String? image;
 
   Question({
     required this.id,
     required this.title,
     required this.question,
     required this.answer,
+    this.references = const [],
+    this.image,
   });
 }