+++ lib/src/screens/catechism_details_screen.dart
@@ -1,3 +1,5 @@
+import 'package:catechism/src/models/tab.dart';
+import 'package:catechism/src/widgets/tab_bar.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -6,7 +8,7 @@ import '../models/questions.dart';
 import '../widgets/app_bar.dart';
 import '../widgets/question_details_item.dart';
 
-class CatechismDetailsScreen extends StatelessWidget {
+class CatechismDetailsScreen extends StatefulWidget {
   /// The CatechismDetailsScreen class is used to create a screen that displays
   /// a QuestionDetailsItem for a given question as well as the catechism
   /// app bar.
@@ -21,6 +23,26 @@ class CatechismDetailsScreen extends StatelessWidget {
   });
 
   @override
+  State<CatechismDetailsScreen> createState() => _CatechismDetailsScreenState();
+}
+
+class _CatechismDetailsScreenState extends State<CatechismDetailsScreen> {
+  late CatechismTabData _selectedTab;
+
+  @override
+  void initState() {
+    super.initState();
+    _selectedTab = widget.configuration.tabs.firstWhere(
+        (element) => element.type == widget.configuration.defaultTabType);
+  }
+
+  void _selectTab(int index) {
+    setState(() {
+      _selectedTab = widget.configuration.tabs[index];
+    });
+  }
+
+  @override
   Widget build(BuildContext context) {
     final questions = Provider.of<Questions>(context).questions;
     final questionId = ModalRoute.of(context)?.settings.arguments as int;
@@ -30,9 +52,8 @@ class CatechismDetailsScreen extends StatelessWidget {
     );
 
     return Scaffold(
-      backgroundColor: Colors.white,
       appBar: CatechismAppBar(
-        titles,
+        widget.titles,
         hasBackButton: true,
       ),
       body: PageView(
@@ -41,10 +62,20 @@ class CatechismDetailsScreen extends StatelessWidget {
         children: questions
             .map((e) => QuestionDetailsItem(
                   question: e,
-                  configuration: configuration,
+                  configuration: widget.configuration,
+                  selectedTab: _selectedTab,
                 ))
             .toList(),
       ),
+      bottomNavigationBar: widget.configuration.showTabBar &&
+              widget.configuration.tabs.length > 1
+          ? CatechismTabBar(
+              questions[questionId - 1],
+              selectTab: _selectTab,
+              selectedTab: _selectedTab,
+              configuration: widget.configuration,
+            )
+          : null,
     );
   }
 }