+++ lib/src/widgets/question_details_item.dart
@@ -1,18 +1,43 @@
+import 'package:catechism/src/models/tab.dart';
 import 'package:flutter/material.dart';
 
 import '../configuration.dart';
 import '../models/question.dart';
+import '../widgets/image_tab.dart';
+import '../widgets/question_tab.dart';
+import '../widgets/scripture_tab.dart';
 
 class QuestionDetailsItem extends StatelessWidget {
   /// The QuestionDetailsItem class is used to display the details of a
   /// question: the title, the number, the question, and the answer.
   final Question question;
   final CatechismConfiguration configuration;
+  final CatechismTabData selectedTab;
 
   QuestionDetailsItem({
     required this.question,
     required this.configuration,
+    required this.selectedTab,
   });
+
+  Widget buildTabContent(context) {
+    switch (selectedTab.type) {
+      case CatechismTabType.image:
+        return ImageTab(question);
+      case CatechismTabType.question:
+        return QuestionTab(
+          question,
+          configuration: configuration,
+        );
+      case CatechismTabType.scripture:
+        return ScriptureTab(
+          question,
+          configuration: configuration,
+        );
+      default:
+        return Container();
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -22,46 +47,7 @@ class QuestionDetailsItem extends StatelessWidget {
           vertical: 10.0,
           horizontal: 35.0,
         ),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            if (configuration.displayQuestionTitleInDetails &&
-                question.title != "")
-              Text(
-                question.title,
-                style: Theme.of(context).textTheme.displaySmall,
-              ),
-            if (configuration.displayQuestionTitleInDetails &&
-                question.title != "")
-              SizedBox(
-                height: 5.0,
-              ),
-            Text(
-              configuration.displayQuestionNumberInDetails
-                  ? '${question.id}. ${question.question}'
-                  : question.question,
-              style: Theme.of(context).textTheme.headlineMedium,
-            ),
-            SizedBox(
-              height: 10.0,
-            ),
-            Divider(
-              thickness: 2.0,
-            ),
-            SizedBox(
-              height: 10.0,
-            ),
-            Text(
-              question.answer,
-              softWrap: true,
-              // textAlign: TextAlign.justify,
-              style: Theme.of(context).textTheme.bodyLarge,
-            ),
-            SizedBox(
-              height: 20.0,
-            ),
-          ],
-        ),
+        child: buildTabContent(context),
       ),
     );
   }