+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3]
+
+### Added
+
+- references for Questions
+- image for Question
+- bottom navigation bar (tabs) on details page
+- tab for showing scripture references/text for a given question
+- tab for showing an image for a given question
+
 ## [0.0.2]
 
 ### Added
@@ -33,6 +43,7 @@ Initial Version
     - displaying the question number in the details screen
     - displaying the question title in the details screen
 
-[unreleased]: https://github.com/fliturgy/catechism/compare/v0.0.2...HEAD
+[unreleased]: https://github.com/fliturgy/catechism/compare/v0.0.3...HEAD
+[0.0.3]: https://github.com/fliturgy/catechism/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/fliturgy/catechism/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/fliturgy/catechism/releases/tag/v0.0.1